anythingllm-auth/requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0